#2985 Add metrics to the native queries underpinning SQL

+++ processing/src/main/java/io/druid/query/DefaultQueryMetrics.java
@@ -160,7 +160,7 @@ public class DefaultQueryMetrics<QueryType extends Query<?>> implements QueryMet
   }
 
   @Override
-  public QueryMetrics<QueryType> reportQueryRows(long rows)
+  public QueryMetrics<QueryType> reportQueryRows(int rows)
   {
     metrics.put("query/rows", rows);
     return this;
